Respect tab stops in unexpand
Also fix a minor bug in the line splitting, and factor that part out so
expand and unexpand arguments are handled the same way.

The algorithm for unexpand is taken from Perl's Text::Tabs, with a minor
tweak to handle trailing spaces that end on a tab stop.

This is a big enough change to bump the minor version number.