Change calling convention to use slurpy & named param
Slurping args makes it natural to call with one or several strings.

Using a named argument for the tab stop option is self-documenting and
also enables expansion (to implement unexpand's --all option in the
future, for example).

Because Raku's for loop flattens a single @array argument, this actually
behaves fine if you pass an array (or several arrays) of strings. So
existing code that doesn't specify the tab stop will work fine.

Still, this is a breaking change in the API. Bump the version number.